+++ core/domain/src/commonMain/kotlin/com/kazemieh/domain/util/balanceImpact.kt
@@ -3,7 +3,7 @@ package com.kazemieh.domain.util
 import com.kazemieh.common.model.Transaction
 import com.kazemieh.common.model.TransactionType
 
-fun Transaction.balanceImpact(): Map<Long, Int> = when (type) {
+fun Transaction.balanceImpact(): Map<Long, Long> = when (type) {
     TransactionType.INCOME -> mapOf(sourceId to amount)
 
     TransactionType.EXPENSE -> mapOf(sourceId to -amount)
